requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1 
Flask-Login==0.5.0
python-dotenv==0.19.1
alembic==1.6.5
psycopg2-binary
Flask-Migrate
Werkzeug==2.0.1
Flask-Mail==0.9.1
itsdangerous>=2.0